[feat] 포트폴리오 비교 기능 구현

See merge request s14-fintech-finance-sub1/S14P21D102!153

+++ frontend/WYE/app/src/main/java/com/d102/wye/domain/model/SavePortfolioParams.kt
@@ -12,6 +12,6 @@ data class SavePortfolioParams(
 
 data class EtfCountItem(
     val ticker: String,
-    val counts: Int  // (investAmount × weight / 100) / currentPrice
+    val counts: Double  // (investAmount × weight / 100) / currentPrice
 )
